requirements.txt
streamlit
pillow
pandas
googletrans==4.0.0-rc1
